Update RepresentationAgreement.md
First batch. There are some significant things to think about not included. One of the big ones is the creation of a fiduciary relationship between the relationship and the DAO. I think it is important to create that level of responsibility. That is, after all, what they should be getting compensated for.