GC03012222 RepresentationAgreement.md (#146)
* Update RepresentationAgreement.md

First batch. There are some significant things to think about not included. One of the big ones is the creation of a fiduciary relationship between the relationship and the DAO. I think it is important to create that level of responsibility. That is, after all, what they should be getting compensated for.

* Update RepresentationAgreement.md

nits

Co-authored-by: z0r0z <92001561+z0r0z@users.noreply.github.com>