GC Merge Jan 9 (#164)
* Update RepresentationAgreement.md

First batch. There are some significant things to think about not included. One of the big ones is the creation of a fiduciary relationship between the relationship and the DAO. I think it is important to create that level of responsibility. That is, after all, what they should be getting compensated for.

* Update RepresentationAgreement.md

nits

* Create Certificate of Independent Legal Advice

Additional liability coverage incoming.

* Update FAQ

ROSS/SHIV/JORDAN - Confirm bold comment on line 17.

* Update FAQ

nits

* Update Certificate of Independent Legal Advice

nits

* Create Limitation of Liability.md

* Create General Liability.md

* Update Limitation of Liability.md

* Update Limitation of Liability.md

* Update Limitation of Liability.md

* Update and rename FAQ to FAQ Setting Up A DAO

Co-authored-by: z0r0z <92001561+z0r0z@users.noreply.github.com>