Update TUNAA.md
I think "Admission" in 2.2 should be reconsidered. There should be optional language for membership denoted by holding an NFT in your wallet and a separate toggle for holding a sufficient number (or any number) of tokens designated by the DAO to provide membership to the DAO. Same can be said for the term "voting tokens" used in Section 2.2. My suggestions in-line are not great, but you get the idea.

2.10 should provide optionality for a ragequit by sale on the secondary market, if we envision DAOs with a free-flowing membership structure to use this document. If we don't can probably ignore all of these comments.

In a couple places, I provided a few permutations of how to describe the DAO acting in accordance with its protocol-level governance mechanisms. Just food for thought. I'd prefer consistency, so pick one you like, if you think they're any good.